Move Random.h to `Utilities/` to match other projects
