+++ src/game/AI/PetAI.cpp
@@ -30,7 +30,7 @@
 #include "Group.h"
 #include "SpellAuraDefines.h"
 #include "Map.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 int PetAI::Permissible(Creature const* creature)
 {
