+++ src/game/Anticheat/WardenAnticheat/WardenModuleMgr.cpp
@@ -28,7 +28,7 @@
 #include "Util.h"
 #include "World.h"
 #include "Log.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "IO/Filesystem/FileSystem.h"
 
 #include <vector>
