+++ src/game/Anticheat/WardenAnticheat/WardenScan.cpp
@@ -31,7 +31,7 @@
 #include "World.h"
 #include "Errors.h"
 #include "Platform/CompilerDefs.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #include <string>
 #include <algorithm>
