+++ src/game/AuctionHouse/AuctionHouseBotMgr.cpp
@@ -7,7 +7,7 @@
 #include "AuctionHouseBotMgr.h"
 #include "Config/Config.h"
 #include "Chat.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 INSTANTIATE_SINGLETON_1(AuctionHouseBotMgr);
