+++ src/game/AutoBroadCastMgr.cpp
@@ -5,7 +5,7 @@
 #include "Policies/SingletonImp.h"
 #include "Util.h"
 #include "AutoBroadCastMgr.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 INSTANTIATE_SINGLETON_1(AutoBroadCastMgr);
