+++ src/game/Battlegrounds/BattleGround.cpp
@@ -36,7 +36,7 @@
 #include "Formulas.h"
 #include "GridNotifiersImpl.h"
 #include "Chat.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "ScriptMgr.h"
 
 namespace MaNGOS
