+++ src/game/Handlers/GroupHandler.cpp
@@ -29,7 +29,7 @@
 #include "ObjectMgr.h"
 #include "Player.h"
 #include "Group.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "SocialMgr.h"
 #include "Util.h"
