+++ src/game/HardcodedEvents.cpp
@@ -10,7 +10,7 @@
 #include "GridSearchers.h"
 #include "world/scourge_invasion.h"
 #include "world/world_event_wareffort.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #include <chrono>
 #include <random>
