+++ src/game/ItemEnchantmentMgr.cpp
@@ -25,7 +25,7 @@
 #include "Log.h"
 #include "ObjectMgr.h"
 #include "ProgressBar.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "World.h"
 
 #include <list>
