+++ src/game/Maps/GridMap.cpp
@@ -26,7 +26,7 @@
 #include "GridMap.h"
 #include "VMapFactory.h"
 #include "MoveMap.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "World.h"
 #include "Policies/SingletonImp.h"
 #include "Util.h"
