+++ src/game/Movement/ConfusedMovementGenerator.cpp
@@ -23,7 +23,7 @@
 #include "MoveSplineInit.h"
 #include "MoveSpline.h"
 #include "PathFinder.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "Transport.h"
 
 template<class T>
