+++ src/game/Movement/FleeingMovementGenerator.cpp
@@ -25,7 +25,7 @@
 #include "PathFinder.h"
 #include "MoveMap.h"
 #include "Map.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #define SEARCH_RANDOM_RADIUS 5.0f
 #define MIN_QUIET_DISTANCE 28.0f
