+++ src/game/Movement/RandomMovementGenerator.cpp
@@ -21,7 +21,7 @@
 #include "Util.h"
 #include "MoveSplineInit.h"
 #include "MoveSpline.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 void RandomMovementGenerator::_setRandomLocation(Creature &creature)
 {
