+++ src/game/Movement/TargetedMovementGenerator.cpp
@@ -30,7 +30,7 @@
 #include "TemporarySummon.h"
 #include "GameObjectAI.h"
 #include "Geometry.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 //-----------------------------------------------//
 template<class T, typename D>
