+++ src/game/Objects/Pet.cpp
@@ -32,7 +32,7 @@
 #include "CreatureAI.h"
 #include "Util.h"
 #include "CharacterDatabaseCache.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 //numbers represent minutes * 100 while happy (you get 100 loyalty points per min while happy)
 uint32 const Pet::LevelUpLoyalty[6] =
