+++ src/game/PlayerBots/BattleBotWaypoints.cpp
@@ -26,7 +26,7 @@
 #include "BattleGroundAV.h"
 #include "BattleGroundWS.h"
 #include "Geometry.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 using namespace Geometry;
