+++ src/game/PlayerBots/PartyBotAI.cpp
@@ -26,7 +26,7 @@
 #include "Spell.h"
 #include "SpellAuras.h"
 #include "Chat.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #include <random>
