+++ src/game/UnitAuraProcHandler.cpp
@@ -31,7 +31,7 @@
 #include "World.h"
 #include "Map.h"
 #include "GridMap.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "ScriptMgr.h"
 
 pAuraProcHandler AuraProcHandler[TOTAL_AURAS] =
