+++ src/game/Weather.cpp
@@ -32,7 +32,7 @@
 #include "WorldPacket.h"
 #include "Log.h"
 #include "ProgressBar.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 INSTANTIATE_SINGLETON_1(WeatherMgr);
