+++ src/realmd/AuthSocket.cpp
@@ -38,7 +38,7 @@
 #include "ClientPatchCache.h"
 #include "Memory/NoDeleter.h"
 #include "Errors.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #include "IO/Networking/Utils.h"
 #include "IO/Networking/AsyncSocket.h"
