+++ src/scripts/PrecompiledHeaders/scriptPCH.h
@@ -30,6 +30,6 @@
 #include "World.h"
 #include "Weather.h"
 #include "TotemAI.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 
 #endif
