+++ src/shared/IO/Networking/DNS.cpp
@@ -2,7 +2,7 @@
 #include "./Internal.h"
 #include "Log.h"
 #include "Errors.h"
-#include "Random.h"
+#include "Utilities/Random.h"
 #include "Util.h"
 #include "IO/SystemErrorToString.h"
